Issue #24 - split of simulation.py into GUI and simulation files
